Add files via upload
Added dist and changed geotiff library as I was compiling them into one file and couldn't find time to fix package.json to load custom builds of components. Added geotiff.js stuff to make it render bicubic and force resampleMethod to be parsed through (not sure if it was needed)
Added components to georaster to allow upscaling of images if not set to nearest neighbour and added extra padding to ensure theres enough data to sample from.